+++ src/server/src/controller/badge.js
@@ -1,3 +1,7 @@
+// SPDX-FileCopyrightText: 2022 SAP SE or an SAP affiliate company and CLA-assistant contributors
+//
+// SPDX-License-Identifier: Apache-2.0
+
 const express = require('express'),
     ejs = require('ejs'),
     fs = require('fs'),