+++ src/server/src/services/org.js
@@ -1,3 +1,7 @@
+// SPDX-FileCopyrightText: 2022 SAP SE or an SAP affiliate company and CLA-assistant contributors
+//
+// SPDX-License-Identifier: Apache-2.0
+
 require('../documents/org');
 const mongoose = require('mongoose');
 const Org = mongoose.model('Org');