+++ src/server/src/services/pullRequest.js
@@ -1,3 +1,7 @@
+// SPDX-FileCopyrightText: 2022 SAP SE or an SAP affiliate company and CLA-assistant contributors
+//
+// SPDX-License-Identifier: Apache-2.0
+
 const url = require('./url')
 const github = require('./github')
 const logger = require('../services/logger')