+++ src/server/src/services/repo.js
@@ -1,3 +1,7 @@
+// SPDX-FileCopyrightText: 2022 SAP SE or an SAP affiliate company and CLA-assistant contributors
+//
+// SPDX-License-Identifier: Apache-2.0
+
 require('../documents/repo')
 const mongoose = require('mongoose')
 const Repo = mongoose.model('Repo')